fix: refactor SummaryAndMap into its own component to avoid initialization problems
